requirements.txt
certifi==2018.11.29
chardet==3.0.4
Click==7.0
clickclick==1.2.2
connexion==2.0.2
Flask==1.0.2
idna==2.7
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.6.0
MarkupSafe==1.1.0
mysql-connector-python==8.0.13
openapi-spec-validator==0.2.4
PyYAML==3.13
requests==2.20.1
six==1.12.0
urllib3==1.24.1
Werkzeug==0.14.1
